Matching Game
This is a matching game which you need to find the extra face on the
screen.
If anything except the correct face is clicked, a message is displayed
saying that the game is over. If the correct face is clicked, all the
currently displayed faces are deleted and a new set of faces is shown
at random positions. Each time a new set of faces is shown there will
be 5 more faces than before, on both the left and the right sides.
There will always be one extra face shown on the left.